Migrate with_no_source_path to BookTest
