+++ tests/rendered_output.rs
@@ -3,17 +3,16 @@ mod dummy_book;
 use crate::dummy_book::{assert_contains_strings, DummyBook};
 
 use anyhow::Context;
-use mdbook::book::Chapter;
 use mdbook::config::Config;
 use mdbook::errors::*;
 use mdbook::utils::fs::write_file;
-use mdbook::{BookItem, MDBook};
+use mdbook::MDBook;
 use pretty_assertions::assert_eq;
 use select::document::Document;
 use select::predicate::{Attr, Class, Name, Predicate};
 use std::ffi::OsStr;
 use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::str::FromStr;
 use tempfile::Builder as TempFileBuilder;
 use walkdir::{DirEntry, WalkDir};
@@ -507,22 +506,4 @@ fn custom_fonts() {
         actual_files(&p.join("book/fonts")),
         &["fonts.css", "myfont.woff"]
     );
-}
-
-#[test]
-fn with_no_source_path() {
-    // Test for a regression where search would fail if source_path is None.
-    let temp = DummyBook::new().build().unwrap();
-    let mut md = MDBook::load(temp.path()).unwrap();
-    let chapter = Chapter {
-        name: "Sample chapter".to_string(),
-        content: "".to_string(),
-        number: None,
-        sub_items: Vec::new(),
-        path: Some(PathBuf::from("sample.html")),
-        source_path: None,
-        parent_names: Vec::new(),
-    };
-    md.book.sections.push(BookItem::Chapter(chapter));
-    md.build().unwrap();
 }
