+++ tests/testsuite/search.rs
@@ -1,8 +1,10 @@
 //! Tests for search support.
 
 use crate::prelude::*;
+use mdbook::book::Chapter;
+use mdbook::BookItem;
 use snapbox::file;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 
 fn read_book_index(root: &Path) -> serde_json::Value {
     let index = root.join("book/searchindex.js");
@@ -103,4 +105,23 @@ fn can_disable_individual_chapters() {
     assert!(!contains("second/"));
     assert!(!contains("first/disable_me.html"));
     assert!(contains("first/keep_me.html"));
+}
+
+// Test for a regression where search would fail if source_path is None.
+// https://github.com/rust-lang/mdBook/pull/2550
+#[test]
+fn with_no_source_path() {
+    let test = BookTest::from_dir("search/reasonable_search_index");
+    let mut book = test.load_book();
+    let chapter = Chapter {
+        name: "Sample chapter".to_string(),
+        content: "".to_string(),
+        number: None,
+        sub_items: Vec::new(),
+        path: Some(PathBuf::from("sample.html")),
+        source_path: None,
+        parent_names: Vec::new(),
+    };
+    book.book.sections.push(BookItem::Chapter(chapter));
+    book.build().unwrap();
 }
